add Bazel build support
For now the main benefit is that it specifies the dependency on the
Vulkan headers, so it can be built at least on some Linux machines
without installing the Vulkan SDK.

Currently this cannot be built a MacOS machine that doesn't
have the Vulkan SDK installed. I'll try to fix that in the future.