Rely on system installation of MoltenVK with Bazel
The previous attempt to use a hermetic copy of it didn't work in all
cases. The dylib was linked properly, but some Vulkan libraries at
runtime access it through the "loader driver interface", which requires
an `icd.d` directory to be present in a particular path.

See
https://vulkan.lunarg.com/doc/view/1.3.211.0/mac/LoaderDriverInterface.html#user-content-driver-discovery-on-macos

So for now we'll just rely on the user to install MoltenVK
(probably via the LunarG Vulkan SDK), but I hope that maybe this is
fixable once https://github.com/bazelbuild/bazel/issues/13930 is fixed.